+++ src/tiri/jit/src/lua.h
@@ -158,9 +158,9 @@ extern void   lua_getfield(lua_State *L, int idx, std::string_view k);
 extern void   lua_rawget(lua_State *L, int idx);
 extern void   lua_rawgeti(lua_State *L, int idx, int n);
 extern void   lua_createtable(lua_State *L, int narr, int nrec);
-extern void   lua_createarray(lua_State *L, uint32_t Length, AET Type, void *Data = nullptr, uint8_t Flags = 0,
+extern void   lua_createarray(lua_State *L, int64_t Length, AET Type, void *Data = nullptr, uint8_t Flags = 0,
    std::string_view StructName = {});
-extern void   lua_createarray(lua_State *L, uint32_t Length, AET Type, void *Data, uint8_t Flags,
+extern void   lua_createarray(lua_State *L, int64_t Length, AET Type, void *Data, uint8_t Flags,
    std::string_view StructName, struct_record *StructDef);
 struct Object;
 extern GCstruct * lua_pushstruct(lua_State *L, struct_record &Def, void *Data = nullptr, uint8_t Flags = 0,
